fix: move boolPromise
