--- packages/shared/src/components/filters/common.tsx
@@ -1,1 +1,0 @@
-export type BooleanPromise = Promise<{ successful: boolean }>;
