+++ packages/shared/src/hooks/useReportPost.ts
@@ -6,11 +6,11 @@ import {
   UNHIDE_POST_MUTATION,
 } from '../graphql/posts';
 import AuthContext from '../contexts/AuthContext';
-import type { BooleanPromise } from '../components/filters/common';
 import { AuthTriggers } from '../lib/auth';
 import { useRequestProtocol } from './useRequestProtocol';
 import { gqlClient } from '../graphql/common';
 import type { ReportReason } from '../report';
+import type { BooleanPromise } from '../lib/func';
 
 type UseReportPostRet = {
   reportPost: (variables: {
